refactor: mv pkg tinypkg -> pkg/tinypkg
