+++ _examples/00same-package/main.go
@@ -8,8 +8,8 @@ import (
 	"m/00same-package/design"
 	"m/fileutil"
 
+	"github.com/podhmo/apikit/pkg/tinypkg"
 	"github.com/podhmo/apikit/resolve"
-	"github.com/podhmo/apikit/tinypkg"
 	"github.com/podhmo/apikit/translate"
 )
