+++ generate/generate.go
@@ -1,0 +1,12 @@
+package generate
+
+type Generator struct {
+	RootDirs    map[string]string
+	RootPackage string
+}
+
+func NewGenerator(rootdir string) *Generator {
+	return &Generator{
+		RootDirs: map[string]string{"": rootdir},
+	}
+}
