+++ pkg/tinypkg/tinypkg.go
