+++ pkg/tinypkg/type_string_test.go
