+++ pkg/tinypkg/walk_test.go
