+++ pkg/tinypkg/write_something.go
