+++ translate/to_runner_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 
 	"github.com/podhmo/apikit/difftest"
+	"github.com/podhmo/apikit/pkg/tinypkg"
 	"github.com/podhmo/apikit/resolve"
-	"github.com/podhmo/apikit/tinypkg"
 )
 
 type Session struct{}
